+++ Assets/Tests/InputSystem/Plugins/UITests.cs
@@ -34,7 +34,7 @@ using UnityEngine.UIElements;
 #pragma warning disable CS0649
 ////TODO: app focus handling
 
-internal class UITests : InputTestFixture
+internal class UITests : CoreTestsFixture
 {
     private struct TestObjects
     {